# back/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RAGU Web API Gateway",
    "summary": "FastAPI gateway for preindexed RAGU graphs.",
    "description": "FastAPI gateway for dataset selection, graph exploration, and question answering over existing RAGU index folders. Live indexing, job queues, and GPU workers are intentionally not exposed.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/datasets": {
      "get": {
        "tags": [
          "Datasets"
        ],
        "summary": "List preindexed datasets",
        "operationId": "list_datasets",
        "parameters": [
          {
            "name": "locale",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "ru",
                "en"
              ],
              "type": "string",
              "description": "Response locale.",
              "default": "ru",
              "title": "Locale"
            },
            "description": "Response locale."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DatasetCard"
                  },
                  "title": "Response List Datasets"
                }
              }
            }
          },
          "404": {
            "description": "Dataset not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets/{dataset_id}": {
      "get": {
        "tags": [
          "Datasets"
        ],
        "summary": "Get dataset details",
        "operationId": "get_dataset",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          },
          {
            "name": "locale",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "ru",
                "en"
              ],
              "type": "string",
              "description": "Response locale.",
              "default": "ru",
              "title": "Locale"
            },
            "description": "Response locale."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetDetail"
                }
              }
            }
          },
          "404": {
            "description": "Dataset not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets/{dataset_id}/graph": {
      "get": {
        "tags": [
          "Graph Explorer"
        ],
        "summary": "Get graph node-link data",
        "operationId": "get_graph",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5000,
              "minimum": 1,
              "default": 500,
              "title": "Limit"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "entity_types",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Types"
            }
          },
          {
            "name": "community_ids",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Community Ids"
            }
          },
          {
            "name": "min_strength",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 1.0,
              "minimum": 0.0,
              "default": 0.0,
              "title": "Min Strength"
            }
          },
          {
            "name": "include_communities",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Include Communities"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphResponse"
                }
              }
            }
          },
          "404": {
            "description": "Dataset or graph resource not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets/{dataset_id}/graph/nodes/{node_id}": {
      "get": {
        "tags": [
          "Graph Explorer"
        ],
        "summary": "Get entity card data",
        "operationId": "get_node",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          },
          {
            "name": "node_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Node Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Dataset or graph resource not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets/{dataset_id}/graph/nodes/{node_id}/neighbors": {
      "get": {
        "tags": [
          "Graph Explorer"
        ],
        "summary": "Get a node neighborhood subgraph",
        "operationId": "get_node_neighbors",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          },
          {
            "name": "node_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Node Id"
            }
          },
          {
            "name": "depth",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 3,
              "minimum": 1,
              "default": 1,
              "title": "Depth"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "min_strength",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 1.0,
              "minimum": 0.0,
              "default": 0.0,
              "title": "Min Strength"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphResponse"
                }
              }
            }
          },
          "404": {
            "description": "Dataset or graph resource not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets/{dataset_id}/graph/communities": {
      "get": {
        "tags": [
          "Graph Explorer"
        ],
        "summary": "List graph communities",
        "operationId": "get_communities",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphCommunitiesResponse"
                }
              }
            }
          },
          "404": {
            "description": "Dataset or graph resource not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets/{dataset_id}/agent/messages": {
      "post": {
        "tags": [
          "Agent"
        ],
        "summary": "Ask a question over a dataset graph",
        "operationId": "create_agent_message",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentResponse"
                }
              }
            }
          },
          "404": {
            "description": "Dataset not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets/{dataset_id}/agent/suggestions": {
      "get": {
        "tags": [
          "Agent"
        ],
        "summary": "Get dataset-specific starter questions",
        "operationId": "get_agent_suggestions",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          },
          {
            "name": "locale",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "ru",
                "en"
              ],
              "type": "string",
              "description": "Response locale.",
              "default": "ru",
              "title": "Locale"
            },
            "description": "Response locale."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestionsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Dataset not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AgentRequest": {
        "properties": {
          "message": {
            "type": "string",
            "minLength": 1,
            "title": "Message",
            "examples": [
              "Who wrote the Norwegian anthem?"
            ]
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            },
            "type": "array",
            "title": "History"
          },
          "engine": {
            "type": "string",
            "enum": [
              "local",
              "global",
              "naive",
              "mix",
              "query_plan"
            ],
            "title": "Engine",
            "default": "local"
          },
          "top_k": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Top K",
            "default": 8
          },
          "rerank": {
            "type": "boolean",
            "title": "Rerank",
            "default": true
          },
          "include_trace": {
            "type": "boolean",
            "title": "Include Trace",
            "default": true
          },
          "locale": {
            "type": "string",
            "enum": [
              "ru",
              "en"
            ],
            "title": "Locale",
            "default": "ru"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "message"
        ],
        "title": "AgentRequest"
      },
      "AgentResponse": {
        "properties": {
          "message": {
            "$ref": "#/components/schemas/AssistantMessage"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "message"
        ],
        "title": "AgentResponse"
      },
...
      "ChatMessage": {
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "user",
              "assistant"
            ],
            "title": "Role",
            "examples": [
              "user"
            ]
          },
          "content": {
            "type": "string",
            "minLength": 1,
            "title": "Content",
            "examples": [
              "Who wrote the Norwegian anthem?"
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "role",
          "content"
        ],
        "title": "ChatMessage"
      },
...
      "DatasetBadge": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label",
            "examples": [
              "model"
            ]
          },
          "value": {
            "type": "string",
            "title": "Value",
            "examples": [
              "meno-lite-7b"
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "label",
          "value"
        ],
        "title": "DatasetBadge"
      },
      "DatasetCard": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "examples": [
              "wiki"
            ]
          },
          "title": {
            "type": "string",
            "title": "Title",
            "examples": [
              "Wikipedia topic slice"
            ]
          },
          "domain": {
            "type": "string",
            "title": "Domain",
            "examples": [
              "Wikipedia"
            ]
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "language": {
            "type": "string",
            "enum": [
              "ru",
              "en",
              "mixed"
            ],
            "title": "Language"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "stats": {
            "$ref": "#/components/schemas/DatasetStats"
          },
          "badges": {
            "items": {
              "$ref": "#/components/schemas/DatasetBadge"
            },
            "type": "array",
            "title": "Badges"
          },
          "preview": {
            "$ref": "#/components/schemas/DatasetPreview"
          },
          "suggested_questions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Suggested Questions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "title",
          "domain",
          "description",
          "language",
          "stats",
          "preview"
        ],
        "title": "DatasetCard"
      },
      "DatasetDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "examples": [
              "wiki"
            ]
          },
          "title": {
            "type": "string",
            "title": "Title",
            "examples": [
              "Wikipedia topic slice"
            ]
          },
          "domain": {
            "type": "string",
            "title": "Domain",
            "examples": [
              "Wikipedia"
            ]
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "language": {
            "type": "string",
            "enum": [
              "ru",
              "en",
              "mixed"
            ],
            "title": "Language"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "stats": {
            "$ref": "#/components/schemas/DatasetStats"
          },
          "badges": {
            "items": {
              "$ref": "#/components/schemas/DatasetBadge"
            },
            "type": "array",
            "title": "Badges"
          },
          "preview": {
            "$ref": "#/components/schemas/DatasetPreview"
          },
          "suggested_questions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Suggested Questions"
          },
          "default_engine": {
            "type": "string",
            "enum": [
              "local",
              "global",
              "naive",
              "mix",
              "query_plan"
            ],
            "title": "Default Engine",
            "default": "local"
          },
          "available_engines": {
            "items": {
              "type": "string",
              "enum": [
                "local",
                "global",
                "naive",
                "mix",
                "query_plan"
              ]
            },
            "type": "array",
            "title": "Available Engines"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "title",
          "domain",
          "description",
          "language",
          "stats",
          "preview",
          "created_at",
          "updated_at"
        ],
        "title": "DatasetDetail"
      },
      "DatasetPreview": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "graph",
            "title": "Kind",
            "default": "graph"
          },
          "node_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Node Count",
            "examples": [
              80
            ]
          },
          "edge_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Edge Count",
            "examples": [
              130
            ]
          },
          "primary_entity_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Primary Entity Types",
            "examples": [
              [
                "PERSON",
                "LAW"
              ]
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "node_count",
          "edge_count"
        ],
        "title": "DatasetPreview"
      },
      "DatasetStats": {
        "properties": {
          "nodes": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Nodes",
            "examples": [
              2400
            ]
          },
          "edges": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Edges",
            "examples": [
              7100
            ]
          },
          "communities": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Communities",
            "examples": [
              38
            ]
          },
          "chunks": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Chunks",
            "examples": [
              320
            ]
          },
          "documents": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Documents",
            "examples": [
              12
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "nodes",
          "edges",
          "communities",
          "chunks",
          "documents"
        ],
        "title": "DatasetStats"
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "examples": [
              "dataset_not_found"
            ]
          },
          "message": {
            "type": "string",
            "title": "Message",
            "examples": [
              "Dataset 'unknown' was not found."
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse"
      },
      "GraphCommunitiesResponse": {
        "properties": {
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id"
          },
          "communities": {
            "items": {
              "$ref": "#/components/schemas/CommunitySummary"
            },
            "type": "array",
            "title": "Communities"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "dataset_id",
          "communities"
        ],
        "title": "GraphCommunitiesResponse"
      },
...
      "GraphResponse": {
        "properties": {
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/GraphNode"
            },
            "type": "array",
            "title": "Nodes"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/GraphEdge"
            },
            "type": "array",
            "title": "Edges"
          },
          "communities": {
            "items": {
              "$ref": "#/components/schemas/CommunitySummary"
            },
            "type": "array",
            "title": "Communities"
          },
          "meta": {
            "$ref": "#/components/schemas/GraphMeta"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "nodes",
          "edges",
          "meta"
        ],
        "title": "GraphResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NodeDetailResponse": {
        "properties": {
          "node": {
            "$ref": "#/components/schemas/GraphNode"
          },
          "incoming_relations": {
            "items": {
              "$ref": "#/components/schemas/NodeRelation"
            },
            "type": "array",
            "title": "Incoming Relations"
          },
          "outgoing_relations": {
            "items": {
              "$ref": "#/components/schemas/NodeRelation"
            },
            "type": "array",
            "title": "Outgoing Relations"
          },
          "provenance_chunks": {
            "items": {
              "$ref": "#/components/schemas/ProvenanceChunk"
            },
            "type": "array",
            "title": "Provenance Chunks"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "node"
        ],
        "title": "NodeDetailResponse"
      },
...
      "SuggestionsResponse": {
        "properties": {
          "dataset_id": {
            "type": "string",
            "title": "Dataset Id"
          },
          "suggestions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Suggestions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "dataset_id",
          "suggestions"
        ],
        "title": "SuggestionsResponse"
      },
...
    }
  },
...
}